changed return type of getUserFavoriteSats to Set<Satellite> from List<Satellite> to avoid return duplicates

+++ src/main/java/com/app/satpoint/repos/UserDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserDAO extends JpaRepository<User, Integer> {
+public interface UserDAO extends JpaRepository<User, Long> {
 
     Optional<User> findUserByUsername(String username);
 }
